Consolidated the portrait and landscape views files into one file and moved the device rotation code
